+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## 3.0
 
+### NEXT
+* `kotlincrypto:secure-random:0.3.2` -> `kotlincrypto.random:crypto-rand:0.5.0`
+    * This fixes key generation in WASM/JS
+* Add `SpecializedSymmetricEncryptionAlgorithm`
+    * This allows `randomKey()` etc to operate on COSE/JWE algorithms
+* Clean up some function signatures:
+    * `SymmetricKey.toJsonWebKey` now returns `KmmResult`
+    * `SymmetricEncryptionAlgorithm.toJweKwAlgorithm` now returns `KmmResult`
+    * `SymmetricEncryptionAlgorithm.toJweEncryptionAlgorithm` removed
+
 ### 3.16.2
 * Lower Android `minSDK` to 21 (5.0 Lollipop) for all modules Except _Supreme_
 * Update AGP to 8.6.1